# Add cached lazy unmarshalling for field sets