**What type of PR is this?**
/kind feature
/wg apply
/sig api-machinery
/priority important-soon
/cc @lavalamp @apelisse @kwiesmueller 

**What this PR does / why we need it**:
This prevents the apiserver from doing any work to re encode a field set that hasn't changed at all during an apply or update. The benefit here is for updates that occur very often but only change one or two fields, like a node heartbeat, etc. This also makes the decoding of the fieldset lazy, which won't give any benefit yet because we always read every fieldset, but in the future we plan to optimize that in structured-merge-diff, and this will make that work possible. Here are some benchmarks:

Before:
```
pkg: k8s.io/apiserver/pkg/endpoints/handlers/fieldmanager
BenchmarkApplyNewObject-12     	    1000	   1674802 ns/op	  500674 B/op	    7167 allocs/op
BenchmarkUpdateNewObject-12    	    1000	   1895343 ns/op	  560238 B/op	    8396 allocs/op
BenchmarkRepeatedUpdate-12     	    2000	    823516 ns/op	  221117 B/op	    2887 allocs/op
```
After
```
pkg: k8s.io/apiserver/pkg/endpoints/handlers/fieldmanager
BenchmarkApplyNewObject-12     	    1000	   1772745 ns/op	  500664 B/op	    7167 allocs/op
BenchmarkUpdateNewObject-12    	    1000	   1823955 ns/op	  558106 B/op	    8362 allocs/op
BenchmarkRepeatedUpdate-12     	    5000	    505358 ns/op	  156538 B/op	    1870 allocs/op
```


**Does this PR introduce a user-facing change?**:
```release-note
NONE
```

**Additional documentation e.g., KEPs (Kubernetes Enhancement Proposals), usage docs, etc.**:
Part of https://github.com/kubernetes/enhancements/pull/1110