@lavalamp 2019-07-22
This caches the structure but not the serialization, right?

I think we need to hook into the serialization process :/

https://github.com/kubernetes-sigs/structured-merge-diff/pull/90