@jennybuckley 2019-07-22
@lavalamp I think we can do both. This is still necessary if we want to avoid deserializing every entry (for example when an update only modifies fields it already manages, we can tell there will be no conflicts early)

/retest