@lavalamp 2019-07-22
Not sure I understand-- if we do something to completely defer
serialization for field sets, don't we get this for free?

On Mon, Jul 22, 2019, 12:51 PM Jennifer Buckley <notifications@github.com>
wrote:

> @lavalamp <https://github.com/lavalamp> I think we can do both. This is
> still necessary if we want to avoid deserializing every entry (for example
> when an update only modifies fields it already manages, we can tell there
> will be no conflicts early)
>
> /retest
>
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub
> <https://github.com/kubernetes/kubernetes/pull/80438?email_source=notifications&email_token=AAE6BFWJT5Z57KBE2DA5BP3QAYFSJA5CNFSM4IF3IWI2YY3PNVWWK3TUL52HS4DFVREXG43VMVBW63LNMVXHJKTDN5WW2ZLOORPWSZGOD2Q6U5A#issuecomment-513927796>,
> or mute the thread
> <https://github.com/notifications/unsubscribe-auth/AAE6BFXDYCJEP6GN6RCABBDQAYFSJANCNFSM4IF3IWIQ>
> .
>