@jennybuckley 2019-07-22
@lavalamp Unless we plan for the interface between the apiserver and smd to be passing in the bytes, I don't think it will defer serialization enough. I had thought that smd#90 was just to expose a faster and more memory efficient serialization/deserialization function for fieldsets, not to defer anything to later