@lavalamp 2019-07-22
Yeah, the part that defers it isn't written. I don't think there's a way to
hook this up to the existing structure at all. The goal is to basically
delete all of the code that fills out the existing structure.

On Mon, Jul 22, 2019 at 2:05 PM Jennifer Buckley <notifications@github.com>
wrote:

> /retest
>
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub
> <https://github.com/kubernetes/kubernetes/pull/80438?email_source=notifications&email_token=AAE6BFW6MYVIKQUFLSUFBGTQAYOH7A5CNFSM4IF3IWI2YY3PNVWWK3TUL52HS4DFVREXG43VMVBW63LNMVXHJKTDN5WW2ZLOORPWSZGOD2RE5YI#issuecomment-513953505>,
> or mute the thread
> <https://github.com/notifications/unsubscribe-auth/AAE6BFV4MC4H3RA4VXCG25DQAYOH7ANCNFSM4IF3IWIQ>
> .
>