@apelisse 2019-07-23
I think we need to:
1. Have Fields be a series of bytes so that it's not deserialized on reads (with a new field name),
2. Use smd#90 to deserialize this series of bytes,
3. Merge this PR so that deserialize when needed, and more importantly, only re-serialize what changed

Technically, with the existing code, READS are still by far the biggest problem we have ...