@jennybuckley 2019-07-24
@apelisse  I agree that we need to do all 3 of those steps but why does the order matter?