@jennybuckley 2019-07-24
>I was also hoping to see the full serialization into a RawExtension/string/[]byte/RawMessage here or in another PR

I think another PR makes more sense