@apelisse 2019-07-24
Order is not super important indeed, @lavalamp has started a PR for the rawextension. 

Returning an error sounds good