@apelisse 2019-07-29
I'm waiting for `Set()` to return an error and I'll LGTM.