@apelisse 2019-08-09
Hey Jenny, I"d love to get this in, do you mind if I follow-up on this?

Thanks